refactor: decompose utils.js for 2 separate files
