+++ src/mock/event-data.js
@@ -1,4 +1,5 @@
-import {getRandomElement, getRandomWeek, getRandomTwoDays, getRandomNumber, getRandomBoolean} from '../utils.js';
+import {getRandomElement, getRandomNumber, getRandomBoolean} from '../utils/common.js';
+import {getRandomWeek, getRandomTwoDays} from '../utils/date.js';
 import {EventType, Destinations, EVENT_DEFAULT} from '../const.js';
 import {generateOfferList} from './offer-data.js';
